Footstep controller and footstep sounds are in a good spot for now
Footsteps have 2 different base terrains: metal and concrete, as well as several layers that can be added on top of that in varying amounts: metal walkway clank sound and metal creak sound.
Finally a water splash amount can be added which filters the base sound a bit and adds a splash sound on top of it.
The controller will probably continue to be tweaked but works decently right now.