New Trap Look
Made the tripwires look more electrical. This was done with two particle systems. Turn up intensity or increase the number of particles emitted to increase "density".
Also feel free to remove the light component on the particles if they are causing too much lag. The are only there for a "flashy" effect